requirements.txt
argparse
pytest-shutil
numpy
scipy
matplotlib
gitpython
joblib
google
google-api-python-client
pandas
Pillow
ezdxf